requirements.txt
pytesseract==0.3.9
Pillow==10.1.0
pyperclip==1.8.2
keyboard==0.13.5
pystray==0.19.5
